refactor some more (get rid of 'CustomKeyboardDispatcher'; weave into 'ShortcutManager')
