+++ src/gwt/src/org/rstudio/core/client/command/KeyboardHelper.java
@@ -16,6 +16,7 @@ package org.rstudio.core.client.command;
 
 import org.rstudio.core.client.BrowseCap;
 
+import com.google.gwt.core.client.JavaScriptObject;
 import com.google.gwt.dom.client.NativeEvent;
 import com.google.gwt.event.dom.client.KeyCodes;
 
@@ -74,4 +75,102 @@ public class KeyboardHelper
          return false;
       }
    }
+   
+   public static int keyCodeFromKeyName(String keyName)
+   {
+      return keyCodeFromKeyName(keyName, KEY_NAME_TO_KEY_CODE_MAP);
+   }
+   
+   private static final native int keyCodeFromKeyName(String keyName, JavaScriptObject map)
+   /*-{
+      return map[keyName] || -1;
+   }-*/;
+   
+   private static final native JavaScriptObject makeKeyCodeToKeyNameMap()
+   /*-{
+      // Map array indices as key codes to corresponding name.
+      var map = new Array(256);
+      
+      map[8] = "backspace";
+      map[9] = "tab";
+      map[12] = "numlock";
+      map[13] = "enter";
+      map[16] = "shift";
+      map[17] = "ctrl";
+      map[18] = "alt";
+      map[19] = "pause";
+      map[20] = "capslock";
+      map[27] = "escape";
+      map[33] = "pageup";
+      map[34] = "pagedown";
+      map[35] = "end";
+      map[36] = "home";
+      map[37] = "left";
+      map[38] = "up";
+      map[39] = "right";
+      map[40] = "down";
+      map[45] = "insert";
+      map[46] = "delete";
+      
+      // Add in numbers 0-9
+      for (var i = 48; i <= 57; i++)
+         map[i] = "" + (i - 48);
+         
+      // Add in letters
+      for (var i = 65; i <= 90; i++)
+         map[i] = String.fromCharCode(i).toLowerCase();
+         
+      map[91] = "left.window";
+      map[92] = "right.window";
+      
+      for (var i = 96; i <= 105; i++)
+         map[i] = "numpad" + (i - 96);
+      
+      map[106] = "*";
+      map[107] = "+";
+      map[109] = "-";
+      
+     // NOTE: This is actually 'decimal point' which is
+     // distinct as a keycode from '.', but probably easier
+     // to just treat them the same.
+      map[110] = "." 
+      
+      map[111] = "/";
+      for (var i = 112; i <= 123; i++)
+         map[i] = "f" + (i - 111);
+      
+      map[144] = "num.lock";
+      map[145] = "scroll.lock";
+      map[186] = ";";
+      map[187] = "=";
+      map[188] = ",";
+      map[189] = "-";
+      map[190] = ".";
+      map[191] = "/";
+      map[192] = "`";
+      map[219] = "[";
+      map[220] = "\\";
+      map[221] = "]";
+      map[222] = "'";
+      
+      return map; 
+   }-*/;
+   
+   private static final native JavaScriptObject makeKeyNameToKeyCodeMap(JavaScriptObject map)
+   /*-{
+      
+      var result = {};
+      for (var key in map) {
+         var value = map[key];
+         result[value] = parseInt(key);
+      }
+      return result;
+      
+   }-*/;
+   
+   private static final JavaScriptObject KEY_CODE_TO_KEY_NAME_MAP =
+         makeKeyCodeToKeyNameMap();
+   
+   private static final JavaScriptObject KEY_NAME_TO_KEY_CODE_MAP =
+         makeKeyNameToKeyCodeMap(KEY_CODE_TO_KEY_NAME_MAP);
 }
