+++ src/gwt/src/org/rstudio/studio/client/RStudioGinjector.java
@@ -19,6 +19,7 @@ import com.google.gwt.inject.client.GinModules;
 import com.google.gwt.inject.client.Ginjector;
 
 import org.rstudio.core.client.command.ShortcutViewer;
+import org.rstudio.core.client.command.UserCommandManager;
 import org.rstudio.core.client.files.filedialog.PathBreadcrumbWidget;
 import org.rstudio.core.client.widget.CaptionWithHelp;
 import org.rstudio.core.client.widget.LocalRepositoriesWidget;
@@ -135,6 +136,7 @@ public interface RStudioGinjector extends Ginjector
    void injectMembers(SetupChunkOptionsPopupPanel panel);
    void injectMembers(SourceSatellite satellite);
    void injectMembers(ModifyKeyboardShortcutsWidget widget);
+   void injectMembers(UserCommandManager manager);
    
    public static final RStudioGinjector INSTANCE = GWT.create(RStudioGinjector.class);
